hardware/ril: Remove ril version checks for qcril second instance.
The ril version check for 2nd instance is put considering DSDS functionality.
In case of non-dsds qcril 2nd instance does not exists. So the ril version
check always become true leading to failure in UUS feature functionality.
But in case of DSDS also Qualcomm Modem is consolidated modem for both qcril
instances and the ril version is same as well. So the ril version check
is removed now.

Change-Id: I9eda46047893358b1fcf80529be0bd43c1981f90
CRs-Fixed: 292309